Smaller changes throughout the app
