+++ Intake/AllergyView/AllergyLLMAssistant.swift
@@ -59,7 +59,7 @@ struct AllergyLLMAssistant: View {
     
     @LLMSessionProvider<LLMOpenAISchema> var session: LLMOpenAISession
     
-    @State var showOnboarding = true
+    @AppStorage(StorageKeys.llmOnboardingComplete) var showOnboarding = true
     @State var greeting = true
     
     @State var allergyItemBox: AllergyItemBox
