+++ Intake/HomeView/Settings/SettingsView.swift
@@ -1,0 +1,55 @@
+//
+// This source file is part of the Stanford LLM on FHIR project
+//
+// SPDX-FileCopyrightText: 2023 Stanford University
+//
+// SPDX-License-Identifier: MIT
+//
+
+import SwiftUI
+
+
+struct SettingsView: View {
+    private enum SettingsDestinations {
+        case resourceSelection
+    }
+
+    @State private var path = NavigationPath()
+    @Environment(\.dismiss) private var dismiss
+
+    var body: some View {
+        NavigationStack(path: $path) {
+            List {
+                resourcesSettings
+            }
+                .navigationTitle("SETTINGS_TITLE")
+                .navigationDestination(for: SettingsDestinations.self) { destination in
+                    navigationDesination(for: destination)
+                }
+                .toolbar {
+                    ToolbarItem(placement: .cancellationAction) {
+                        Button("FHIR_RESOURCES_CHAT_CANCEL") {
+                            dismiss()
+                        }
+                    }
+                }
+        }
+    }
+
+    private var resourcesSettings: some View {
+        Section("Resource Selection") {
+            NavigationLink(value: SettingsDestinations.resourceSelection) {
+                Text("Resource Selection")
+            }
+        }
+    }
+
+    private func navigationDesination(for destination: SettingsDestinations) -> some View {
+        Group {
+            switch destination {
+            case .resourceSelection:
+                ResourceSelection()
+            }
+        }
+    }
+}
