+++ Intake/MedicationView/MedicationLLMAssistant.swift
@@ -20,7 +20,7 @@ struct MedicationLLMAssistant: View {
     
     @LLMSessionProvider<LLMOpenAISchema> var session: LLMOpenAISession
     
-    @State var showOnboarding = true
+    @AppStorage(StorageKeys.llmOnboardingComplete) var showOnboarding = true
     @State var greeting = true
     
     var body: some View {
