+++ Intake/Resources/Localizable.xcstrings
@@ -148,9 +148,6 @@
     "Back" : {
 
     },
-    "Chat" : {
-
-    },
     "Chief Complaint:" : {
 
     },
