+++ Intake/SocialHistoryView/MenstrualHistory.swift
@@ -21,43 +21,38 @@ struct SocialHistoryQuestionView: View {
     
     
     var body: some View {
-        NavigationView {
-            VStack {
-                if data.generalData.sex == "Female" {
-                    Form {
-                        Section(header: Text("Menstrual Information").foregroundColor(.gray)) {
-                            @Bindable var data = data
-                            DatePicker("Last period's start date", selection: $startDate, in: ...Date(), displayedComponents: .date)
-                                .datePickerStyle(DefaultDatePickerStyle())
-                            
-                            DatePicker("Last period's end date", selection: $endDate, in: ...Date(), displayedComponents: .date)
-                                .datePickerStyle(DefaultDatePickerStyle())
-                        }
-
-                        Section(header: Text("Additional Symptoms").foregroundColor(.gray)) {
-                            @Bindable var data = data
-                            TextField("Ex: Heavy bleeding on second day, fatigue...", text: $additionalDetails)
-                        }
+        VStack {
+            if data.generalData.sex == "Female" {
+                Form {
+                    Section(header: Text("Menstrual Information").foregroundColor(.gray)) {
+                        @Bindable var data = data
+                        DatePicker("Last period's start date", selection: $startDate, in: ...Date(), displayedComponents: .date)
+                            .datePickerStyle(DefaultDatePickerStyle())
+                        
+                        DatePicker("Last period's end date", selection: $endDate, in: ...Date(), displayedComponents: .date)
+                            .datePickerStyle(DefaultDatePickerStyle())
                     }
-                    .navigationTitle("Social History")
-                    .task {
-                        startDate = data.menstrualHistory.startDate
-                        endDate = data.menstrualHistory.endDate
-                        additionalDetails = data.menstrualHistory.additionalDetails
+                    
+                    Section(header: Text("Additional Symptoms").foregroundColor(.gray)) {
+                        @Bindable var data = data
+                        TextField("Ex: Heavy bleeding on second day, fatigue...", text: $additionalDetails)
                     }
-                    /*.task {
-                        fetchHealthKitData()
-                    }*/
-                    .onDisappear {
-                        data.menstrualHistory = MenstrualHistoryItem(startDate: startDate, endDate: endDate, additionalDetails: additionalDetails)
-                    }
-                    if FeatureFlags.skipToScrollable {
-                        SubmitButton(nextView: NavigationViews.pdfs)
-                            .padding()
-                    } else {
-                        SubmitButton(nextView: NavigationViews.smoking)
-                            .padding()
-                    }
+                }
+                .navigationTitle("Menstrual History")
+                .task {
+                    startDate = data.menstrualHistory.startDate
+                    endDate = data.menstrualHistory.endDate
+                    additionalDetails = data.menstrualHistory.additionalDetails
+                }
+                .onDisappear {
+                    data.menstrualHistory = MenstrualHistoryItem(startDate: startDate, endDate: endDate, additionalDetails: additionalDetails)
+                }
+                if FeatureFlags.skipToScrollable {
+                    SubmitButton(nextView: NavigationViews.pdfs)
+                        .padding()
+                } else {
+                    SubmitButton(nextView: NavigationViews.smoking)
+                        .padding()
                 }
             }
         }
