+++ Intake/SurgeryView/SurgeryLLMAssistant.swift
@@ -56,7 +56,7 @@ struct SurgeryLLMAssistant: View {
     
     @LLMSessionProvider<LLMOpenAISchema> var session: LLMOpenAISession
     
-    @State var showOnboarding = true
+    @AppStorage(StorageKeys.llmOnboardingComplete) var showOnboarding = true
     @State var greeting = true
     
     @State var surgeryItemBox: SurgeryItemBox
